[Perils and Princesses] Semi-New SHeet
Since 22Jun2025 the "Perils & Princesses" sheet has been unable to access any images. The only response to the problem was the suggestion that actual locations should be used. Last month this was done and nothing has changed. Without concrete evidence, I am guessing its the ampersand in the folder name