Small change to uncouple weapons & companions from using the same fieldset.
Little concerned about this one since it will blow away any companions on existing sheets. I'm not sure if there's a way to push a migration as part of a change, also not sure how roll20 works with character sheet updates; if you automatically update, or if you only get the latest on creating a new campaign.

We could have an sheet:opened event that checks if any weapon fieldsets have companion data but that's a bit of clutter, and will mostly not do anything. Not looked into the specifics of how to write that one.