Add Land of Eem character sheet
Initial commit of the Land of Eem character sheet.

css and html point to my repository as a source of all the images (https://github.com/blaze-sanecki/LandOfEem). I can upload those assets here and update the links if needed.

I tried my best to follow the guidelines, but feel free to let me know if anything needs to be fixed/changed.